allow click on another tile and replace
this is a completion to the prev commit -
which denied a specialist from being removed if 20 tiles were worked.
but it prohibited a tile switch , unless another one was removed instead...
so this will trigger a tile change to true if there are 20 tiles.
in the set workingplot - the code checks above 20 and removed the worser tile - so it works as a combined system